requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
commandlines==0.4.1
cycler==0.11.0
filetype==1.0.9
Flask==2.0.2
fonttools==4.28.5
hsh==1.1.0
idna==3.3
imageio==2.13.5
itsdangerous==2.0.1
Jinja2==3.0.3
kiwisolver==1.3.2
llvmlite==0.37.0
MarkupSafe==2.0.1
matplotlib==3.5.1
networkx==2.6.3
numba==0.54.1
numpy==1.20.3
packaging==21.3
Pillow==8.4.0
PyMatting==1.1.4
pyparsing==3.0.6
python-dateutil==2.8.2
PyWavelets==1.2.0
rembg==1.0.27
requests==2.26.0
scikit-image==0.19.1
scipy==1.7.3
six==1.16.0
tifffile==2021.11.2
torch==1.10.1
torchvision==0.11.2
tqdm==4.62.3
typing_extensions==4.0.1
urllib3==1.26.7
waitress==2.0.0
Werkzeug==2.0.2
